mac80211: Skip punctured channel validation for GHz
Avoid channel validation of sub channels that are punctured.
5G 240Mhz channel is 320-80 channel. Avoid validating punctured
80Mhz channel in the 320Mhz bandwidth

+++ backports-6.1-9a0dddfb3/include/net/cfg80211.h
@@ -5222,7 +5222,7 @@ struct ieee80211_iface_combination {
 	 * @radar_detect_widths:
 	 * bitmap of channel widths supported for radar detection
 	 */
-	u8 radar_detect_widths;
+	u32 radar_detect_widths;
 
 	/**
 	 * @radar_detect_regions:
